Improve Hover Editor & Insider build support
My original PR for adding file menus to the plugin incorrectly
passed a "leaf" argument when opening the file menu, which
causes certain menu items to misbehave, especially in newer
versions of Obsidian.  (It also prevents Hover Editor from
adding a menu item to open the targeted file.)

This tiny patch resolves these issues by removing the leaf
argument, which according to the Obsidian devs should not
be passed unless the file in question is *in* the leaf
(which in Recent Files' case, it never is).